--- src/main/java/net/detalk/api/controller/v1/response/GetTagResponse.java
@@ -1,5 +1,0 @@
-package net.detalk.api.controller.v1.response;
-
-public record GetTagResponse(String name) {
-
-}
