+++ src/main/java/net/detalk/api/tag/controller/TagController.java
@@ -1,9 +1,9 @@
-package net.detalk.api.controller.v1;
+package net.detalk.api.tag.controller;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import net.detalk.api.controller.v1.response.GetTagResponse;
-import net.detalk.api.service.TagService;
+import net.detalk.api.tag.controller.response.GetTagResponse;
+import net.detalk.api.tag.service.TagService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
