+++ src/main/java/net/detalk/api/tag/controller/response/GetTagResponse.java
@@ -1,0 +1,3 @@
+package net.detalk.api.tag.controller.response;
+
+public record GetTagResponse(String name) {}
