+++ src/main/java/net/detalk/api/tag/repository/TagRepository.java
@@ -1,0 +1,15 @@
+package net.detalk.api.tag.repository;
+
+import java.util.List;
+import java.util.Optional;
+import net.detalk.api.tag.domain.Tag;
+
+public interface TagRepository {
+
+    Optional<Tag> findByName(String name);
+
+    Tag save(Tag tag);
+
+    List<Tag> findAll();
+
+}
