+++ src/main/java/net/detalk/api/tag/service/TagService.java
@@ -1,10 +1,10 @@
-package net.detalk.api.service;
+package net.detalk.api.tag.service;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import net.detalk.api.controller.v1.response.GetTagResponse;
-import net.detalk.api.domain.Tag;
-import net.detalk.api.repository.TagRepository;
+import net.detalk.api.tag.controller.response.GetTagResponse;
+import net.detalk.api.tag.domain.Tag;
+import net.detalk.api.tag.repository.TagRepository;
 import org.springframework.stereotype.Service;
 
 @RequiredArgsConstructor
