+++ src/test/java/net/detalk/api/service/ProductPostServiceTest.java
@@ -29,7 +29,7 @@ import net.detalk.api.product.domain.Product;
 import net.detalk.api.product.domain.ProductLink;
 import net.detalk.api.post.domain.ProductPost;
 import net.detalk.api.post.domain.ProductPostSnapshot;
-import net.detalk.api.domain.Tag;
+import net.detalk.api.tag.domain.Tag;
 import net.detalk.api.domain.exception.DuplicateCreatePostException;
 import net.detalk.api.domain.exception.InvalidRecommendCountRequest;
 import net.detalk.api.mock.FakeTimeHolder;
@@ -43,6 +43,7 @@ import net.detalk.api.support.CursorPageData;
 import net.detalk.api.support.TimeHolder;
 import net.detalk.api.support.UUIDGenerator;
 import net.detalk.api.support.error.ApiException;
+import net.detalk.api.tag.service.TagService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
